API: Add commission charge, fix return url, fix empty winners

+++ backend/bidfall/payment.py
@@ -83,7 +83,7 @@ def cancel_payment(payment_id: str) -> dict:
         raise
 
 
-def capture_payment(payment_id: str, amount: float = None):
+def capture_payment(payment_id: str, amount: Decimal = None):
     capture_data = {}
     if amount is not None:
         capture_data["amount"] = {
